requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
gunicorn
pyspark==3.5.1
boto3
python-dotenv
requests
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
# Security & Authentication
cryptography
Authlib
pytest
pydantic==2.11.1
passlib[bcrypt]==1.7.4
python-multipart
